requirements.txt
grpc-stubs==1.24.8
grpcio==1.44.0
grpcio-tools==1.44.0
mypy-protobuf==3.2.0
protobuf==3.19.4
six==1.16.0
types-protobuf==3.19.12
yapf==0.32.0
